backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.1
pydantic==2.5.0
pydantic-settings==2.1.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
pytesseract==0.3.10
openpyxl==3.1.2

# Additional utilities
aiofiles==23.2.1
python-magic==0.4.27
